Reorganizes Codecs
